Added apply_slave_role_relations()
Given a list of relations of which we are slaves, we apply each one and
include the returned object on our foreign fields.

For now, we include the actual object, and assume DBIC will do the right
thing, but this only works with recent DBIx::Class releases
(after October 2010 *I think*).

Future versions could probably extract the fields specified in the
relationship conditions, but thats basically rewritting
_resolve_condition() here...

Also there is something that we don't support cleanly right now: if, when
we apply() on the master of our slave relation, the row ends up deleted,
we don't propagate this to us. We probably should, but I need more real
world usage to understand the tradeoffs of each option.

Signed-off-by: Pedro Melo <melo@simplicidade.org>