Note that we need to check for the present of Row objects in some cases
In some places we need to copy some fields from SRC to DEST. Right now,
we are assuming the type of some of those SRC/DEST to be a Row object or
a HashRef.

We should extrat this logic to a new helper function that deals with all
combinations of HashRef, Row, etc.

Signed-off-by: Pedro Melo <melo@simplicidade.org>