Add support for special field __ID:
If present, it will be replaced with the proper PK fields. This allows us
to target the PK of a source without knowing the field name.

We do need to know the PK cardinality though. If the PK has multiple
columns, we need to pass a ArrayRef with the exact number of values.

Signed-off-by: Pedro Melo <melo@simplicidade.org>