On a slave => master cenario, use slave FKs to complete master data
If we need to follow a relation from a slave to a master (like a
belongs_to) and the data given for the master side doesn't have enough
information to find a unique row, we use any FKs we can find on the slave
side.

The logic is that, if we want to update a field on the master side of the
current master object, we should only need the fields we need to update,
and apply() would use the FKs from the current object to do find it.

Signed-off-by: Pedro Melo <melo@simplicidade.org>