Add support for __ACTION => 'CREATE'
If forces the creation of a row, even if we could find one via one of the
unique constraints.

For a sinlge table apply, this doesn't make a lot of sense, because it
would probably thrown unique constraints exceptions, but it is useful
to break a slave row from its master table.

Assume that Emails belongs to Users, then you can change a email to a new
user like this:

    $email->apply({
      user => { __ACTION => 'CREATE', login => 'login_new' }
    });

Without the forced __ACTION => 'CREATE', the login of the current user
would be updated to 'login_new'.

Signed-off-by: Pedro Melo <melo@simplicidade.org>