Add support for fake_master roles:
To be used on belongs_to relationships to a slave table that has his
pk as fk on the master side.

The usual use case is when X has_many Y but one Y in particular is the
active one.

In those situations, I like to have the active PK as a FK in X. The
op sequence is:

 * create X;
 * create Y, using X pk as fk;
 * update X to set the Y pk as fk.

Signed-off-by: Pedro Melo <melo@simplicidade.org>